requirements.txt
selenium==4.15.0
google-generativeai==0.3.0
python-dotenv==1.0.0
pillow==10.0.1
webdriver-manager==3.8.6